Create separate test suite for MovieList
