+++ app/src/main/java/sleepyweasel/purplefluffernutter/MovieContentProvider.java
@@ -1,0 +1,4 @@
+package sleepyweasel.purplefluffernutter;
+
+public class MovieContentProvider {
+}
